Initialize skip list with NULL value.
Initialize the skip list with a NULL value. The horizontal and vertical nodes
both point to that value to save space. Also add a new member of the
'struct node_t' struct to represent the level this node is located on. And
finally I removed the bottom member of the 'struct skip_list_t' because I
believe it is not needed.